Massive refactoring of Quartz to: - Define each tag in their own class. - Allow extension functions to additional responsibilities to other classes - Migrate from hardcoded tag filters in events to the Tag's parser and assemble functions. - Migrate hardcoded event.create to builders that use extension functions - Restructures threading infrastructure for NIP-10 - Decouple the event signing from the Event building functions via event templates - Create classes to represent Tags and TagArrays and use extension functions to add domain-related methods to the tag array of each nip. - Uses external functions on event template builders to better point to which functions and which tags can be used in which event kinds. - Separates Event kinds in packages inside each nip. - Improves support for NIP-89 - Correctly establishes which imeta params can be used in each nip (video, picture, files) - Decouples the iMeta builder from any nip. - Fixes mute list word and user removal when inserted from a different client. - Migrates the Account class to avoiding having to build each Event inside of it

+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/actions/NewPollClosing.kt
@@ -52,7 +52,7 @@ fun NewPollClosing(pollViewModel: NewPostViewModel) {
     pollViewModel.isValidClosedAt.value = true
     if (text.isNotEmpty()) {
         try {
-            val int = text.toInt()
+            val int = text.toLong()
             if (int < 0) {
                 pollViewModel.isValidClosedAt.value = false
             } else {
